+++ src/session/redis/mod.rs
@@ -4,16 +4,16 @@ use crate::prelude::*;
 use redis::Client;
 
 impl SessionManager for Client {
-    fn insert(&self, id: u64, key: String) -> Result<()> {
+    fn insert(&self, id: u32, key: String) -> Result<()> {
         todo!()
     }
-    fn insert_for(&self, id: u64, key: String, time: Duration) -> Result<()> {
+    fn insert_for(&self, id: u32, key: String, time: Duration) -> Result<()> {
         todo!()
     }
-    fn remove(&self, id: u64) -> Result<()> {
+    fn remove(&self, id: u32) -> Result<()> {
         todo!()
     }
-    fn get(&self, id: u64) -> Option<String> {
+    fn get(&self, id: u32) -> Option<String> {
         todo!()
     }
     fn flush(&self) -> Result<()> {
